Cascading deletes (#11)
This PR ensures that deletes to a Household resource will cascade to its members as well. This non-cascading anomaly was found earlier when the tests are written and it was discovered that member objects remain orphaned in the database even when its household is deleted. Since the only interaction is via the /households/ endpoint, it makes little sense not to allow deletes to cascade since there will otherwise be no way for the orphaned member to be recreated (due to an enforce UNIQUE constraint applied on the name of a FamilyMember).

* ensure tests fail and show that deletes to a household is not being cascaded down to its members

* set FamilyMember instance to delete when a Household is deleted